requirements.txt
paho-mqtt==1.5.0
websocket-client==0.57.0
python-socketio==4.6.0
